feat(test_fetcher): support file download from series
