+++ tests/e2e/test_fetcher/__init__.py
@@ -3,9 +3,13 @@ __all__ = [
     "fetch_metric_values",
     "fetch_series_values",
     "fetch_files",
+    "fetch_file_series",
 ]
 
 from .attribute_values import fetch_attribute_values
-from .files import fetch_files
+from .files import (
+    fetch_file_series,
+    fetch_files,
+)
 from .metric import fetch_metric_values
 from .series import fetch_series_values
